requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
openai>=1.3.0
pyyaml>=6.0
schedule>=1.2.0
feedparser>=6.0.10
